Move systemd classes to Yast2 namespace
