+++ library/network/src/modules/NetworkService.rb
@@ -42,7 +42,7 @@
 # The network.service alias link obsoletes the old master switch in
 # /etc/sysconfig/network/config:NETWORKMANAGER (until openSUSE-12.2).
 require "yast"
-require "yast2/systemd_service"
+require "yast2/systemd/service"
 
 module Yast
   class NetworkServiceClass < Module
@@ -193,7 +193,7 @@ module Yast
         @current_name = DEFAULT_BACKEND
         log.info "Running in installer/AutoYaST, use default: #{@current_name}"
       else
-        service = SystemdService.find("network")
+        service = Yast2::Systemd::Service.find("network")
         @current_name = BACKENDS.invert[service.name] if service
       end
