+++ library/runlevel/test/test_helper.rb
@@ -1,4 +1,4 @@
-# We need the stubs used for SystemdService tests
+# We need the stubs used for Yast2::Systemd::Service tests
 require_relative "../../systemd/test/test_helper"
 
 require_relative "../../../test/test_helper.rb"
