+++ library/systemd/examples/service_widget.rb
@@ -7,8 +7,8 @@ require "yast2/compound_service"
 def service
   return @service if @service
 
-  service1 = Yast2::SystemService.find("cups.service")
-  service2 = Yast2::SystemService.find("dbus.service")
+  service1 = Yast2::SystemService.find!("cups.service")
+  service2 = Yast2::SystemService.find!("dbus.service")
 
   # Yast2::ServiceWidget can be used with both, a Yast2::SystemService or
   # a Yast2::CompoundService
