+++ library/systemd/src/lib/yast2/systemd.rb
@@ -1,0 +1,7 @@
+require "yast2/systemd/unit"
+require "yast2/systemd/target"
+require "yast2/systemd/service"
+require "yast2/systemd/socket"
+
+# Systemd control API
+module Systemd; end
