+++ library/systemd/test/Makefile.am
@@ -1,18 +1,2 @@
-TESTS = \
-  systemctl_test.rb \
-  systemd_unit_test.rb \
-  systemd_socket_test.rb \
-  systemd_target_test.rb \
-  systemd_service_test.rb \
-  yast2/compound_service_test.rb \
-  yast2/service_widget_test.rb \
-  yast2/system_service_test.rb \
-  yast2/system_service_test.rb \
-  yast2/systemd_socket_finder_test.rb
-
-TEST_EXTENSIONS = .rb
-RB_LOG_COMPILER = rspec
-VERBOSE = 1
-
-EXTRA_DIST = $(TESTS) test_helper.rb
+EXTRA_DIST = test_helper.rb
