+++ library/systemd/test/test_helper.rb
@@ -1,10 +1,11 @@
 require_relative "../../../test/test_helper.rb"
 
-require "yast2/systemd_unit"
-require "yast2/systemd_service"
-require "yast2/systemd_socket"
-require "yast2/systemd_target"
+require "yast2/systemd/unit"
+require "yast2/systemd/service"
+require "yast2/systemd/socket"
+require "yast2/systemd/target"
 
+SYSTEMD_DATA_PATH = File.expand_path("data", __dir__)
 
 # Find a Term in given content
 #
@@ -38,7 +39,7 @@ module SystemctlStubs
   end
 
   def stub_execute(success: true)
-    allow(Yast::Systemctl).to receive(:execute).and_return(
+    allow(Yast2::Systemctl).to receive(:execute).and_return(
       OpenStruct.new(
         stdout: "success",
         stderr: (success ? "" : "failure"),
@@ -48,7 +49,7 @@ module SystemctlStubs
   end
 
   def stub_socket_unit_files
-    allow(Yast::Systemctl).to receive(:list_unit_files).and_return(<<LIST
+    allow(Yast2::Systemctl).to receive(:list_unit_files).and_return(<<LIST
 iscsid.socket                disabled
 avahi-daemon.socket          enabled
 cups.socket                  enabled
@@ -59,7 +60,7 @@ LIST
   end
 
   def stub_service_unit_files
-    allow(Yast::Systemctl).to receive(:list_unit_files).and_return(<<LIST
+    allow(Yast2::Systemctl).to receive(:list_unit_files).and_return(<<LIST
 single.service                             masked
 smartd.service                             disabled
 smb.service                                disabled
@@ -71,7 +72,7 @@ LIST
   end
 
   def stub_target_unit_files
-    allow(Yast::Systemctl).to receive(:list_unit_files).and_return(<<LIST
+    allow(Yast2::Systemctl).to receive(:list_unit_files).and_return(<<LIST
 graphical.target          enabled
 halt.target               disabled
 hibernate.target          static
@@ -85,7 +86,7 @@ LIST
   end
 
   def stub_service_units
-    allow(Yast::Systemctl).to receive(:list_units).and_return(<<LIST
+    allow(Yast2::Systemctl).to receive(:list_units).and_return(<<LIST
 rsyslog.service                       loaded active   running System Logging Service
 scsidev.service                       not-found inactive dead    scsidev.service
 sendmail.service                      not-found inactive dead    sendmail.service
@@ -97,7 +98,7 @@ LIST
   end
 
   def stub_socket_units
-    allow(Yast::Systemctl).to receive(:list_units).and_return(<<LIST
+    allow(Yast2::Systemctl).to receive(:list_units).and_return(<<LIST
 iscsid.socket                loaded active   listening Open-iSCSI iscsid Socket
 avahi-daemon.socket          loaded active   running   Avahi mDNS/DNS-SD Stack Activation Socket
 cups.socket                  loaded inactive dead      CUPS Printing Service Sockets
@@ -110,7 +111,7 @@ LIST
   end
 
   def stub_target_units
-    allow(Yast::Systemctl).to receive(:list_units).and_return(<<LIST
+    allow(Yast2::Systemctl).to receive(:list_units).and_return(<<LIST
 getty.target           loaded active   active Login Prompts
 graphical.target       loaded inactive dead   Graphical Interface
 local-fs-pre.target    loaded active   active Local File Systems (Pre)
@@ -126,7 +127,7 @@ end
 
 module SystemdUnitStubs
   def stub_unit_command(success: true)
-    allow_any_instance_of(Yast::SystemdUnit)
+    allow_any_instance_of(Yast2::Systemd::Unit)
       .to receive(:command)
       .and_return(
         OpenStruct.new(
@@ -154,7 +155,7 @@ module SystemdSocketStubs
     stub_unit_command
     stub_systemctl(:socket)
     properties = load_socket_properties(socket)
-    allow_any_instance_of(Yast::SystemdUnit::Properties)
+    allow_any_instance_of(Yast2::Systemd::Unit::Properties)
       .to receive(:load_systemd_properties)
       .and_return(properties)
   end
@@ -168,7 +169,7 @@ module SystemdServiceStubs
     stub_unit_command
     stub_systemctl(:service)
     properties = load_service_properties(service)
-    allow_any_instance_of(Yast::SystemdUnit::Properties)
+    allow_any_instance_of(Yast2::Systemd::Unit::Properties)
       .to receive(:load_systemd_properties)
       .and_return(properties)
   end
@@ -190,7 +191,7 @@ module SystemdTargetStubs
     stub_unit_command
     stub_systemctl(:target)
     properties = load_target_properties(target)
-    allow_any_instance_of(Yast::SystemdUnit::Properties)
+    allow_any_instance_of(Yast2::Systemd::Unit::Properties)
       .to receive(:load_systemd_properties)
       .and_return(properties)
   end
