+++ library/systemd/test/yast2/systemd_service_test.rb
@@ -1,9 +1,9 @@
 #!/usr/bin/env rspec
 
-require_relative "test_helper"
+require_relative "../test_helper"
 
-module Yast
-  describe SystemdService do
+module Yast2
+  describe Systemd::Service do
     include SystemdServiceStubs
 
     before do
@@ -13,8 +13,8 @@ module Yast
     describe ".find" do
       it "returns the service unit object specified in parameter" do
         ["sshd", "sshd.service"].each do |service_name|
-          service = SystemdService.find(service_name)
-          expect(service).to be_a(SystemdUnit)
+          service = Systemd::Service.find(service_name)
+          expect(service).to be_a(Systemd::Unit)
           expect(service.unit_type).to eq("service")
           expect(service.unit_name).to eq("sshd")
         end
@@ -25,13 +25,13 @@ module Yast
           properties = OpenStruct.new(
             stdout: "", stderr: "Unit unknown.service could not be found.", exit: 1
           )
-          allow_any_instance_of(Yast::SystemdUnit::Properties)
+          allow_any_instance_of(Systemd::Unit::Properties)
             .to receive(:load_systemd_properties)
             .and_return(properties)
         end
 
         it "returns nil" do
-          service = SystemdService.find("another")
+          service = Systemd::Service.find("another")
           expect(service).to be_nil
         end
       end
@@ -40,8 +40,8 @@ module Yast
     describe ".build" do
       it "returns the service unit object specified in parameter" do
         ["sshd", "sshd.service"].each do |service_name|
-          service = SystemdService.build(service_name)
-          expect(service).to be_a(SystemdUnit)
+          service = Systemd::Service.build(service_name)
+          expect(service).to be_a(Systemd::Unit)
           expect(service.unit_type).to eq("service")
           expect(service.unit_name).to eq("sshd")
         end
@@ -49,22 +49,22 @@ module Yast
 
       it "returns a service instance even if the real service does not exist" do
         stub_services(service: "unknown")
-        service = SystemdService.build("unknown")
+        service = Systemd::Service.build("unknown")
         expect(service.name).to eq("unknown")
       end
     end
 
     describe ".find!" do
       it "returns the service unit object specified in parameter" do
-        service = SystemdService.find("sshd")
-        expect(service).to be_a(SystemdUnit)
+        service = Systemd::Service.find("sshd")
+        expect(service).to be_a(Systemd::Unit)
         expect(service.unit_type).to eq("service")
         expect(service.unit_name).to eq("sshd")
       end
 
-      it "raises SystemdServiceNotFound error if unit does not exist" do
+      it "raises Systemd::ServiceNotFound error if unit does not exist" do
         stub_services(service: "unknown")
-        expect { SystemdService.find!("unknown") }.to raise_error(SystemdServiceNotFound)
+        expect { Systemd::Service.find!("unknown") }.to raise_error(Systemd::ServiceNotFound)
       end
     end
 
@@ -73,20 +73,20 @@ module Yast
       let(:apparmor_double) { double("Service", name: "apparmor") }
       let(:cups_double) { double("Service", name: "cups") }
       let(:systemctl_stdout) do
-        File.read(File.join(__dir__, "data", "apparmor_and_cups_properties"))
+        File.read(File.join(SYSTEMD_DATA_PATH, "apparmor_and_cups_properties"))
       end
 
       before do
-        allow(Yast::Systemctl).to receive(:execute).with(
+        allow(Yast2::Systemctl).to receive(:execute).with(
           "show  --property=Id,MainPID,Description,LoadState,ActiveState,SubState,UnitFileState," \
           "FragmentPath,CanReload apparmor.service cups.service"
         ).and_return(systemctl_show)
-        allow(SystemdService).to receive(:find).with("apparmor", {}).and_return(apparmor_double)
-        allow(SystemdService).to receive(:find).with("cups", {}).and_return(cups_double)
+        allow(Systemd::Service).to receive(:find).with("apparmor", {}).and_return(apparmor_double)
+        allow(Systemd::Service).to receive(:find).with("cups", {}).and_return(cups_double)
       end
 
       it "returns the list of services" do
-        services = SystemdService.find_many(["apparmor", "cups"])
+        services = Systemd::Service.find_many(["apparmor", "cups"])
         expect(services).to contain_exactly(
           an_object_having_attributes("name" => "apparmor"),
           an_object_having_attributes("name" => "cups")
@@ -97,14 +97,14 @@ module Yast
         let(:not_found_double) { double("Service", name: "cups", not_found?: true) }
 
         before do
-          allow(Yast::SystemdService).to receive(:new).and_call_original
-          allow(Yast::SystemdService).to receive(:new)
+          allow(Systemd::Service).to receive(:new).and_call_original
+          allow(Systemd::Service).to receive(:new)
             .with("cups.service", anything, anything)
             .and_return(not_found_double)
         end
 
         it "does not include the not found service" do
-          services = SystemdService.find_many(["apparmor", "cups"])
+          services = Systemd::Service.find_many(["apparmor", "cups"])
           expect(services.map(&:name)).to eq(["apparmor"])
         end
       end
@@ -113,7 +113,7 @@ module Yast
         let(:systemctl_show) { OpenStruct.new(stdout: "", stderr: "", exit: 1) }
 
         it "retrieve services information in a one-by-one basis" do
-          expect(SystemdService.find_many(["apparmor", "cups"]))
+          expect(Systemd::Service.find_many(["apparmor", "cups"]))
             .to eq([apparmor_double, cups_double])
         end
       end
@@ -122,7 +122,7 @@ module Yast
         let(:systemctl_show) { OpenStruct.new(stdout: "", stderr: "error", exit: 1) }
 
         it "retrieve services information in a one-by-one basis" do
-          expect(SystemdService.find_many(["apparmor", "cups"]))
+          expect(Systemd::Service.find_many(["apparmor", "cups"]))
             .to eq([apparmor_double, cups_double])
         end
       end
@@ -130,7 +130,7 @@ module Yast
 
     describe ".all" do
       it "returns all supported services found" do
-        services = SystemdService.all
+        services = Systemd::Service.all
         expect(services).to be_a(Array)
         expect(services).not_to be_empty
         services.each { |s| expect(s.unit_type).to eq("service") }
@@ -139,7 +139,7 @@ module Yast
 
     describe "#running?" do
       it "returns true if the service is running" do
-        service = SystemdService.find "sshd"
+        service = Systemd::Service.find "sshd"
         expect(service).to respond_to(:running?)
         expect(service.running?).to eq(true)
       end
@@ -147,7 +147,7 @@ module Yast
 
     describe "#pid" do
       it "returns the pid of the running service" do
-        service = SystemdService.find("sshd")
+        service = Systemd::Service.find("sshd")
         expect(service).to respond_to(:pid)
         expect(service.pid).not_to be_empty
       end
@@ -156,22 +156,22 @@ module Yast
     context "Start a service on the installation system" do
       it "starts a service with a specialized inst-sys helper if available" do
         allow(File).to receive(:exist?).with("/bin/service_start").and_return(true)
-        service = SystemdService.find("sshd")
-        allow(SCR).to receive(:Execute).and_return("stderr" => "", "stdout" => "", "exit" => 0)
-        expect(service).not_to receive(:command) # SystemdUnit#command
+        service = Systemd::Service.find("sshd")
+        allow(Yast::SCR).to receive(:Execute).and_return("stderr" => "", "stdout" => "", "exit" => 0)
+        expect(service).not_to receive(:command) # Systemd::Unit#command
         expect(service.start).to eq(true)
       end
     end
 
     context "Restart a service on the installation system" do
       it "restarts a service with a specialized inst-sys helper if available" do
-        allow_any_instance_of(SystemdService).to receive(:sleep).and_return(1)
+        allow_any_instance_of(Systemd::Service).to receive(:sleep).and_return(1)
         allow(File).to receive(:exist?).with("/bin/service_start").and_return(true)
-        service = SystemdService.find("sshd")
-        allow(SCR).to receive(:Execute).and_return("stderr" => "", "stdout" => "", "exit" => 0)
+        service = Systemd::Service.find("sshd")
+        allow(Yast::SCR).to receive(:Execute).and_return("stderr" => "", "stdout" => "", "exit" => 0)
         expect(service).to receive(:stop).ordered.and_call_original
         expect(service).to receive(:start).ordered.and_call_original
-        expect(service).not_to receive(:command) # SystemdUnit#command
+        expect(service).not_to receive(:command) # Systemd::Unit#command
         expect(service.restart).to eq(true)
       end
     end
@@ -179,26 +179,26 @@ module Yast
     context "Stop a service on the installation system" do
       it "stops a service with a specialized inst-sys helper" do
         allow(File).to receive(:exist?).with("/bin/service_start").and_return(true)
-        service = SystemdService.find("sshd")
-        allow(SCR).to receive(:Execute).and_return("stderr" => "", "stdout" => "", "exit" => 0)
-        expect(service).not_to receive(:command) # SystemdUnit#command
+        service = Systemd::Service.find("sshd")
+        allow(Yast::SCR).to receive(:Execute).and_return("stderr" => "", "stdout" => "", "exit" => 0)
+        expect(service).not_to receive(:command) # Systemd::Unit#command
         expect(service.stop).to eq(true)
       end
     end
 
     describe "#socket" do
-      subject(:service) { SystemdService.find(service_name) }
+      subject(:service) { Systemd::Service.find(service_name) }
       let(:service_name) { "sshd" }
-      let(:socket) { instance_double(SystemdSocket) }
+      let(:socket) { instance_double(Systemd::Socket) }
 
       it "returns the socket for the service" do
-        expect(SystemdSocket).to receive(:for_service).with(service_name)
+        expect(Systemd::Socket).to receive(:for_service).with(service_name)
           .and_return(socket)
         expect(service.socket).to eq(socket)
       end
 
       it "asks for the socket only once" do
-        expect(SystemdSocket).to receive(:for_service).with(service_name)
+        expect(Systemd::Socket).to receive(:for_service).with(service_name)
           .and_return(socket).once
         expect(service.socket).to eq(socket)
         expect(service.socket).to eq(socket)
@@ -206,7 +206,7 @@ module Yast
 
       context "when no associated socket is found" do
         before do
-          allow(SystemdSocket).to receive(:for_service).with(service_name)
+          allow(Systemd::Socket).to receive(:for_service).with(service_name)
             .and_return(nil)
         end
 
@@ -217,14 +217,14 @@ module Yast
     end
 
     describe "#socket?" do
-      subject(:service) { SystemdService.find("sshd") }
+      subject(:service) { Systemd::Service.find("sshd") }
 
       before do
         allow(service).to receive(:socket).and_return(socket)
       end
 
       context "when there is an associated socket" do
-        let(:socket) { instance_double(SystemdSocket) }
+        let(:socket) { instance_double(Systemd::Socket) }
 
         it "returns true" do
           expect(service.socket?).to eq(true)
