+++ library/systemd/test/yast2/systemd_unit_test.rb
@@ -1,14 +1,14 @@
 #!/usr/bin/env rspec
 
-require_relative "test_helper"
+require_relative "../test_helper"
 
-module Yast
-  describe SystemdUnit do
+module Yast2
+  describe Systemd::Unit do
     include SystemdSocketStubs
     include SystemdServiceStubs
 
     def trigger_reloading_properties(command)
-      unit = SystemdUnit.new("new.socket")
+      unit = Systemd::Unit.new("new.socket")
       properties = unit.properties
       unit.send(command)
       expect(unit.properties).not_to equal(properties)
@@ -27,32 +27,32 @@ module Yast
       describe "#properties" do
         context "Unit found" do
           it "returns struct with restricted installation properties" do
-            allow_any_instance_of(SystemdUnit).to receive(:command)
+            allow_any_instance_of(Systemd::Unit).to receive(:command)
               .with("is-enabled sshd.service").and_return(
                 OpenStruct.new("stderr" => "", "stdout" => "enabled", "exit" => 0)
               )
-            unit = SystemdUnit.new("sshd.service")
-            expect(unit.properties).to be_a(SystemdUnit::InstallationProperties)
+            unit = Systemd::Unit.new("sshd.service")
+            expect(unit.properties).to be_a(Systemd::Unit::InstallationProperties)
             expect(unit.properties.not_found?).to eq(false)
           end
 
           describe "#enabled?" do
             it "returns true if service is enabled" do
-              allow_any_instance_of(SystemdUnit).to receive(:command)
+              allow_any_instance_of(Systemd::Unit).to receive(:command)
                 .with("is-enabled sshd.service").and_return(
                   OpenStruct.new("stderr" => "", "stdout" => "enabled", "exit" => 0)
                 )
-              unit = SystemdUnit.new("sshd.service")
+              unit = Systemd::Unit.new("sshd.service")
               expect(unit.enabled?).to eq(true)
             end
 
             it "returns false if service is disabled" do
               stub_unit_command(success: false)
-              allow_any_instance_of(SystemdUnit).to receive(:command)
+              allow_any_instance_of(Systemd::Unit).to receive(:command)
                 .with("is-enabled sshd.service").and_return(
                   OpenStruct.new("stderr" => "", "stdout" => "disabled", "exit" => 1)
                 )
-              unit = SystemdUnit.new("sshd.service")
+              unit = Systemd::Unit.new("sshd.service")
               expect(unit.enabled?).to eq(false)
             end
           end
@@ -62,8 +62,8 @@ module Yast
           it "returns struct with restricted installation properties" do
             stub_services(service: "unknown")
             stub_unit_command(success: false)
-            unit = SystemdUnit.new("unknown.service")
-            expect(unit.properties).to be_a(SystemdUnit::InstallationProperties)
+            unit = Systemd::Unit.new("unknown.service")
+            expect(unit.properties).to be_a(Systemd::Unit::InstallationProperties)
             expect(unit.properties.not_found?).to eq(true)
           end
         end
@@ -72,12 +72,12 @@ module Yast
 
     describe "#properties" do
       it "always returns struct including default properties" do
-        unit = SystemdUnit.new("iscsi.socket")
-        expect(unit.properties.to_h.keys).to include(*SystemdUnit::DEFAULT_PROPMAP.keys)
+        unit = Systemd::Unit.new("iscsi.socket")
+        expect(unit.properties.to_h.keys).to include(*Systemd::Unit::DEFAULT_PROPMAP.keys)
       end
 
       it "provides status properties methods" do
-        unit = SystemdUnit.new("iscsid.socket")
+        unit = Systemd::Unit.new("iscsid.socket")
         expect(unit.properties[:enabled?]).not_to be_nil
         expect(unit.properties[:active?]).not_to be_nil
         expect(unit.properties[:loaded?]).not_to be_nil
@@ -91,7 +91,7 @@ module Yast
       end
 
       it "delegates the status properties onto the unit object" do
-        unit = SystemdUnit.new("iscsid.socket")
+        unit = Systemd::Unit.new("iscsid.socket")
         expect(unit).to respond_to(:enabled?)
         expect(unit).to respond_to(:active?)
         expect(unit).to respond_to(:loaded?)
@@ -101,32 +101,32 @@ module Yast
     end
 
     describe ".new" do
-      it "creates a new SystemdUnit instance with unit name and type parsed from first parameter" do
+      it "creates a new Systemd::Unit instance with unit name and type parsed from first parameter" do
         instance = nil
-        expect { instance = SystemdUnit.new("random.socket") }.not_to raise_error
+        expect { instance = Systemd::Unit.new("random.socket") }.not_to raise_error
         expect(instance.unit_name).to eq("random")
         expect(instance.unit_type).to eq("socket")
       end
 
       it "correctly parses a name with many dots" do
         instance = nil
-        expect { instance = SystemdUnit.new("dbus-org.freedesktop.hostname1.service") }.not_to raise_error
+        expect { instance = Systemd::Unit.new("dbus-org.freedesktop.hostname1.service") }.not_to raise_error
         expect(instance.unit_name).to eq("dbus-org.freedesktop.hostname1")
         expect(instance.unit_type).to eq("service")
       end
 
       it "raises an exception if an incomplete unit name is passed" do
-        expect { SystemdUnit.new("sshd") }.to raise_error(RuntimeError)
+        expect { Systemd::Unit.new("sshd") }.to raise_error(RuntimeError)
       end
 
       it "allows to create supported units" do
-        expect { SystemdUnit.new("my.socket")      }.not_to raise_error
-        expect { SystemdUnit.new("default.target") }.not_to raise_error
-        expect { SystemdUnit.new("sshd.service")   }.not_to raise_error
+        expect { Systemd::Unit.new("my.socket")      }.not_to raise_error
+        expect { Systemd::Unit.new("default.target") }.not_to raise_error
+        expect { Systemd::Unit.new("sshd.service")   }.not_to raise_error
       end
 
       it "accepts parameters to extend the default properties" do
-        unit = SystemdUnit.new("iscsid.socket", requires: "Requires", wants: "Wants")
+        unit = Systemd::Unit.new("iscsid.socket", requires: "Requires", wants: "Wants")
         expect(unit.properties.wants).not_to be_nil
         expect(unit.properties.requires).not_to be_nil
       end
@@ -135,13 +135,13 @@ module Yast
     describe "#stop" do
       it "returns true if unit has been stopped" do
         stub_unit_command
-        unit = SystemdUnit.new("my.socket")
+        unit = Systemd::Unit.new("my.socket")
         expect(unit.stop).to eq(true)
       end
 
       it "returns false if failed" do
         stub_unit_command(success: false)
-        unit = SystemdUnit.new("my.socket")
+        unit = Systemd::Unit.new("my.socket")
         expect(unit.stop).to eq(false)
         expect(unit.error).not_to be_empty
       end
@@ -154,13 +154,13 @@ module Yast
     describe "#start" do
       it "returns true if starts (activates) the unit" do
         stub_unit_command(success: true)
-        unit = SystemdUnit.new("my.socket")
+        unit = Systemd::Unit.new("my.socket")
         expect(unit.start).to eq(true)
       end
 
       it "returns false if failed to start the unit" do
         stub_unit_command(success: false)
-        unit = SystemdUnit.new("my.socket")
+        unit = Systemd::Unit.new("my.socket")
         expect(unit.start).to eq(false)
         expect(unit.error).not_to be_empty
       end
@@ -173,13 +173,13 @@ module Yast
     describe "#enable" do
       it "returns true if the unit has been enabled successfully" do
         stub_unit_command(success: true)
-        unit = SystemdUnit.new("your.socket")
+        unit = Systemd::Unit.new("your.socket")
         expect(unit.enable).to eq(true)
       end
 
       it "returns false if unit fails" do
         stub_unit_command(success: false)
-        unit = SystemdUnit.new("your.socket")
+        unit = Systemd::Unit.new("your.socket")
         expect(unit.enable).to eq(false)
       end
 
@@ -191,13 +191,13 @@ module Yast
     describe "#disable" do
       it "returns true if the unit has been disabled successfully" do
         stub_unit_command(success: true)
-        unit = SystemdUnit.new("your.socket")
+        unit = Systemd::Unit.new("your.socket")
         expect(unit.disable).to eq(true)
       end
 
       it "returns false if unit disabling fails" do
         stub_unit_command(success: false)
-        unit = SystemdUnit.new("your.socket")
+        unit = Systemd::Unit.new("your.socket")
         expect(unit.disable).to eq(false)
       end
 
@@ -208,14 +208,14 @@ module Yast
 
     describe "#show" do
       it "always returns new unit properties object" do
-        unit = SystemdUnit.new("startrek.socket")
+        unit = Systemd::Unit.new("startrek.socket")
         expect(unit.show).not_to equal(unit.show)
       end
     end
 
     describe "#refresh!" do
       it "rewrites and returns the properties instance variable" do
-        unit = SystemdUnit.new("your.socket")
+        unit = Systemd::Unit.new("your.socket")
         properties = unit.properties
         expect(unit.refresh!).not_to equal(properties)
       end
@@ -223,65 +223,65 @@ module Yast
 
     describe "#restart" do
       it "returns true if unit has been restarted" do
-        unit = SystemdUnit.new("sshd.service")
+        unit = Systemd::Unit.new("sshd.service")
         expect(unit.restart).to eq(true)
       end
 
       it "returns false if the restart fails" do
         stub_unit_command(success: false)
-        unit = SystemdUnit.new("sshd.service")
+        unit = Systemd::Unit.new("sshd.service")
         expect(unit.restart).to eq(false)
       end
     end
 
     describe "#try_restart" do
       it "returns true if the unit has been restarted" do
-        unit = SystemdUnit.new("sshd.service")
+        unit = Systemd::Unit.new("sshd.service")
         expect(unit.try_restart).to eq(true)
       end
 
       it "returns false if the try_restart fails" do
         stub_unit_command(success: false)
-        unit = SystemdUnit.new("sshd.service")
+        unit = Systemd::Unit.new("sshd.service")
         expect(unit.try_restart).to eq(false)
       end
     end
 
     describe "#reload" do
       it "returns true if the unit has been reloaded" do
-        unit = SystemdUnit.new("sshd.service")
+        unit = Systemd::Unit.new("sshd.service")
         expect(unit.reload).to eq(true)
       end
 
       it "returns false if the reload fails" do
         stub_unit_command(success: false)
-        unit = SystemdUnit.new("sshd.service")
+        unit = Systemd::Unit.new("sshd.service")
         expect(unit.reload).to eq(false)
       end
     end
 
     describe "#reload_or_restart" do
       it "returns true if the unit has been reloaded or restarted" do
-        unit = SystemdUnit.new("sshd.service")
+        unit = Systemd::Unit.new("sshd.service")
         expect(unit.reload_or_restart).to eq(true)
       end
 
       it "returns false if the reload_or_restart action fails" do
         stub_unit_command(success: false)
-        unit = SystemdUnit.new("sshd.service")
+        unit = Systemd::Unit.new("sshd.service")
         expect(unit.reload_or_restart).to eq(false)
       end
     end
 
     describe "#reload_or_try_restart" do
       it "returns true if the unit has been reload_or_try_restarted" do
-        unit = SystemdUnit.new("sshd.service")
+        unit = Systemd::Unit.new("sshd.service")
         expect(unit.reload_or_try_restart).to eq(true)
       end
 
       it "returns false if the reload_or_try_restart action fails" do
         stub_unit_command(success: false)
-        unit = SystemdUnit.new("sshd.service")
+        unit = Systemd::Unit.new("sshd.service")
         expect(unit.reload_or_try_restart).to eq(false)
       end
     end
@@ -289,7 +289,7 @@ module Yast
     describe "#error" do
       it "returns empty string if the unit commands succeed" do
         stub_unit_command(success: true)
-        unit = SystemdUnit.new("your.socket")
+        unit = Systemd::Unit.new("your.socket")
         unit.start
         expect(unit.error).to be_empty
 
@@ -305,7 +305,7 @@ module Yast
 
       it "returns error string if the unit commands fail" do
         stub_unit_command(success: false)
-        unit = SystemdUnit.new("your.socket")
+        unit = Systemd::Unit.new("your.socket")
         unit.start
         expect(unit.error).not_to be_empty
